refactor: ProductService 인터페이스 분리 (DI/테스트 용이성 향상)
