+++ src/main/java/com/zeebra/domain/product/controller/ProductRestController.java
@@ -1,13 +1,11 @@
 package com.zeebra.domain.product.controller;
 
-import com.zeebra.domain.product.dto.GetProductListRequest;
 import com.zeebra.domain.product.dto.ProductDetailResponse;
-import com.zeebra.domain.product.dto.ProductListResponse;
 import com.zeebra.domain.product.service.ProductService;
+import com.zeebra.domain.product.service.ProductServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
